better store management
