+++ web/greenwave-ui/src/lib/stores/core.js
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+
+// Core data stores
+export const junctions = writable([]);
+export const desiredSpeed = writable(40.0);
